backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
pillow==10.4.0
numpy==1.26.4
aiofiles==24.1.0
python-multipart==0.0.20

# Basic dependencies first, complex ones commented out for now
# We'll add these back once basic API is working

# openimageio==3.0.10.1  # Build is failing, will try alternative approach

# PyTorch and AI dependencies
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Training and data processing
albumentations==1.3.1
# wandb==0.16.0  # Keep this commented for now
psutil==5.9.6
tqdm==4.66.1
PyYAML==6.0.1
opencv-python==4.8.1.78
scipy==1.11.4

# Image and video processing
imageio==2.31.5
# ffmpeg-python==0.2.0